Organize documentation by task (users vs developers/extendors)
Hopefully this makes it easier for people who just want to know how to
write a configuration file.

This is almost completely a mechanical reorginazation with very minor
text edits. The must substantial edits were to the introduction portion
to modernize it (remove specific references to Zope Corporation and
Python 2.3).

A future PR might seek to move much of the separate function/method
documentation into docstrings and use autodoc so that its up to date.