using-logging doc: factor example confs into files and enable testing
This makes the file shorter (but harder to read on its own since you
can't see the example configs anymore), but it avoids repeating
ourself when we load the configuration file in the examples.

Reading from a file is a better example anyway since that's what
usually happens.

Make test_readme.py also test this.