Use manuel to run doctest files.
We can't use the codeblock plugin for the using-logging test, instead
resorting to manually execing the captured source. This is because
stdout is not captured by a codeblock, so logging would be configured
to go to the *real* stdout, not one captured by the next doctest we
do (where we do want to use the doctest syntax) and our output
wouldn't get seen. This was maddening to figure out.

You can't use the capture plugin to capture literal includes, so we
still have the chdir hacks.

Overall it's not a whole lot of a win, at least right now.

I changed some of the text strings that we log to make them a bit more
descriptive. This was part of debuging too, but I think they're
better.

I ran into a problem doing 'python -m ZConfig.tests.test_readme' that
I added a workaround for.