Make tox -e coverage work
Not installing test dependencies such as manuel was causing problems.

I've also tried to fix coverage reporting not to include the entire
Python stdlib, which works fine for `coverage report`, but nose's xcover
plugin still outputs garbage.

I'm not a fan of plugins that stand between me and coverage.py and break
things.